+++ src/compas_fab/ghpython/components/Cf_ConfigMerge/code.py
@@ -1,0 +1,15 @@
+"""
+Merge two configurations.
+
+COMPAS FAB v0.22.0
+"""
+from ghpythonlib.componentbase import executingcomponent as component
+
+
+class ConfigMerge(component):
+    def RunScript(self, config_a, config_b):
+
+        if config_a and config_b:
+            return config_a.merged(config_b)
+
+        return None